Refactor history DB row projection
